requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
langchain-chroma 
pymupdf